apdiffscript/pyproject.toml
[project]
name = "apdiffscript"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "scriptworker",
    "semver",
    "taskcluster",
    "GitPython",
    "scriptworker_common",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.uv.sources]
"scriptworker_common" = { workspace = true }

[build-system]
requires =  ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.sources]
"src/" = ""

